#69 feat(GameOverController): add remove 2 obstacles before continue game
Fix laser flashing material issues.
Rewrited dependencies of lasers.
