+++ HippieFall/Sourse/Biome/Cyber/BiomeCyberConfig.cs
@@ -9,10 +9,10 @@ namespace HippieFall.Biomes
         {       
             TunnelsObstacle = new List<string>
             {
-                C_TunnelObstaclePath.CYBER_CIRCLE_MOVING,
+                /*C_TunnelObstaclePath.CYBER_CIRCLE_MOVING,
                 C_TunnelObstaclePath.CYBER_PERFORATED_WALL,
                 C_TunnelObstaclePath.CYBER_HALF_WALL,
-                C_TunnelObstaclePath.CYBER_FAN,
+                C_TunnelObstaclePath.CYBER_FAN,*/
                 C_TunnelObstaclePath.CYBER_LASER_FASTENERS
             };  
             Gate = "res://HippieFall/Sourse/Biome/Cyber/Gate/Gate.tscn";
